Changed: Extract selectionPostFixer() to own file.
